UI Changes
Changed the Hide Button Color to red so that user can differentiate visually.

Life Form Color adding is to the right instead of the bottom.

Changed Drop Down to UI so its more clean.

Centered Assign Taxa information, it was on the left side of the screen before

Changed history page table color to gray  from white so that it's easier to look at.

Changed the history table to start at the top of the screen.

Separate the File Buttons from the Tool Buttons with "Graphing Tools" title.

Fixed the text width of the about page so that it's easier to read.

Center and change text layout on home page.

Added image layout for tutorial page.

Center the text on tutorial page

Changed image on tutorial page to new graph

Moved the Update Password container more to the center on the Profile Page